requirements.txt
Django==5.1.5
djangorestframework==3.15.2
numpy==2.1.3
faiss-cpu==1.9.0.post1
torch==2.5.1
facenet-pytorch==2.6.0
opencv-python==4.10.0.84
Pillow==11.0.0
requests==2.32.3
customtkinter==5.2.2
cryptography==44.0.0
